// src/app/core/enums/common.enums.ts
export enum Key {
  ESCAPE = 'Escape',
  SPACE = ' ',
  ARROW_LEFT = 'ArrowLeft',
  ARROW_UP = 'ArrowUp',
  ARROW_RIGHT = 'ArrowRight',
  ARROW_DOWN = 'ArrowDown',
}

export enum ColorsEnum {
  NONE = '',
  CYAN = 'cyan',
  BLUE = 'blue',
  ORANGE = 'oranage',
  YELLOW = 'yellow',
  GREEN = 'green',
  PURPLE = 'purple',
  RED = 'red',
}